delete throneAPI old project
