--- ThroneAPI/App_Start/RouteConfig.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.Routing;
-using System.Web.Mvc;
-
-namespace ThroneAPI
-{
-    public class RouteConfig
-    {
-        public static void RegisterRoutes(RouteCollection routes)
-                {
-                    routes.IgnoreRoute("{resource}.axd/{*pathInfo}");
-
-                    routes.MapRoute(
-                        name: "Default",
-                        url: "{controller}/{action}/{id}",
-                        defaults: new { controller = "Home", action = "Index", id = UrlParameter.Optional }
-                    );
-                }
-    }
-}
