--- ThroneAPI/Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="ThroneAPI.WebApiApplication" Language="C#" %>
