--- ThroneAPI/Models/HouseDto.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace ThroneAPI.Models
-{
-    public class HouseDto
-    {
-        public int ID { get; set; }
-        public string Name { get; set; }
-        public int NumberOfUnits { get; set; }
-        // TODO public LinkedList<Character> Housers { get; set; }
-    }
-}
