+++ Throne_of_ZZ.sln
@@ -21,12 +21,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataAccessLayer", "StubData
 		{6847E0CF-0A3B-4EFD-854E-5AE0899BFF47} = {6847E0CF-0A3B-4EFD-854E-5AE0899BFF47}
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ThroneAPI", "ThroneAPI\ThroneAPI.csproj", "{EB04283B-F828-4981-A308-67FE2BE2BFDB}"
-	ProjectSection(ProjectDependencies) = postProject
-		{7E2F1D91-EF02-4628-B512-30E190194949} = {7E2F1D91-EF02-4628-B512-30E190194949}
-		{6847E0CF-0A3B-4EFD-854E-5AE0899BFF47} = {6847E0CF-0A3B-4EFD-854E-5AE0899BFF47}
-	EndProjectSection
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZZThrone", "ZZThrone\ZZThrone.csproj", "{14B2BE28-2A8E-4E44-94B3-80435883427C}"
 EndProject
 Global
@@ -51,10 +45,6 @@ Global
 		{ADE744E1-E54D-4878-9B65-E4A6FE17811C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ADE744E1-E54D-4878-9B65-E4A6FE17811C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{ADE744E1-E54D-4878-9B65-E4A6FE17811C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EB04283B-F828-4981-A308-67FE2BE2BFDB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EB04283B-F828-4981-A308-67FE2BE2BFDB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EB04283B-F828-4981-A308-67FE2BE2BFDB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EB04283B-F828-4981-A308-67FE2BE2BFDB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{14B2BE28-2A8E-4E44-94B3-80435883427C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{14B2BE28-2A8E-4E44-94B3-80435883427C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{14B2BE28-2A8E-4E44-94B3-80435883427C}.Release|Any CPU.ActiveCfg = Release|Any CPU
