+++ ZZThrone/Models/CharacterDto.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
-namespace ThroneAPI.Models
+namespace ZZThrone.Models
 {
     public class CharacterDto
     {
