+++ ZZThrone/Models/TerritoryDto.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
-namespace ThroneAPI.Models
+namespace ZZThrone.Models
 {
     public class Territory
     {
